backend/pyproject.toml
[project]
name = "llmezi-backend"
version = "0.1.0"
description = "A backend service for llmezi application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic<2.0.0,>=1.15.2",
    # Pin bcrypt until passlib supports the latest version
    "bcrypt==4.0.1",
    "fastapi[all]<1.0.0,>=0.115.12",
    "passlib[bcrypt]<2.0.0,>=1.7.4",
    "psycopg[binary]<4.0.0,>=3.2.7",
    "pyjwt<3.0.0,>=2.10.1",
    "sentry-sdk[fastapi]<3.0.0,>=2.27.0",
    "sqlmodel<1.0.0,>=0.0.24",
    "strawberry-graphql[fastapi]<1.0.0,>=0.266.0",
]

[dependency-groups]
dev = [
    "pre-commit<5.0.0,>=4.2.0",
    "ruff<1.0.0,>=0.11.7",
]

[tool.ruff]
line-length = 100
exclude = ["alembic"]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"

[tool.ruff.lint]
extend-select = ["I"]
